github-repo-access-plugin/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
PyGithub==2.1.1
python-dotenv==1.0.0
httpx==0.25.1